Rework into 'desired' vs 'dekorate' tests.
The dekorate variety makes it possible to start now, and evaluate what
the dekorate approach will look like and whether it works.

The desired variety shows what we would like the tests to look like. If
too many changes will be needed to dekorate, we will remove this folder.
Otherwise, we will migrate tests to the dekorate approach.

Add dependencies on the generated images.